mk: Improve STRIP_DEBUG.
The first issue was that on platforms where strip was set to a noop the
command would succeed but the subsequent mv would fail.  Improve this by
actually checking for the output file.

The second issue is that not all strip commands support -g, so we need
an explicit gstrip tool for GNU strip (or equivalents, the FreeBSD
"elftoolchain" supports -g too).

Thirdly, we need a way to support skipping either entire packages or
just a subset of files within a package.  Some binaries (e.g. go) are
broken by being stripped.

While here simplify and tidy the install-strip-debug target, as well as
improve performance by skipping symlinks.

+++ mk/install/install.mk
@@ -179,7 +179,7 @@ _INSTALL_ALL_TARGETS+=		plist
 .if ${_PKGSRC_USE_CTF} == "yes"
 _INSTALL_ALL_TARGETS+=		install-ctf
 .endif
-.if !empty(STRIP_DEBUG:M[Yy][Ee][Ss])
+.if ${STRIP_DEBUG:Uno:tl} == "yes" && ${STRIP_DEBUG_SUPPORTED:Uyes:tl} == "yes"
 _INSTALL_ALL_TARGETS+=		install-strip-debug
 .endif
 _INSTALL_ALL_TARGETS+=		install-doc-handling
@@ -382,16 +382,20 @@ install-ctf: plist
 .PHONY: install-strip-debug
 install-strip-debug: plist
 	@${STEP_MSG} "Automatic stripping of debug information"
-	${RUN}${CAT} ${_PLIST_NOKEYWORDS} \
-	| ${SED} -e 's|^|${DESTDIR}${PREFIX}/|' \
-	| while read f; do \
-		tmp_f="$${f}.XXX"; \
-		if ${STRIP} -g -o "$${tmp_f}" "$${f}" 2> /dev/null; then \
-			[ ! -f "$${f}" ] || \
-			    ${MV} "$${tmp_f}" "$${f}"; \
-		else \
-			${RM} -f "$${tmp_f}"; \
-		fi \
+	${RUN}cd ${DESTDIR:Q}${PREFIX:Q};				\
+	${CAT} ${_PLIST_NOKEYWORDS} | while read f; do			\
+		[ ! -h "$${f}" ] || continue;				\
+		case "$${f}" in						\
+		${STRIP_FILES_SKIP:@p@${p}) continue;;@}		\
+		*) ;;							\
+		esac;							\
+		tmp_f="$${f}.XXX";					\
+		if ${STRIP_DBG} -o "$${tmp_f}" "$${f}" 2>/dev/null; then \
+			if [ -f "$${tmp_f}" -a -f "$${f}" ]; then	\
+				${MV} "$${tmp_f}" "$${f}";		\
+			fi;						\
+		fi;							\
+		${RM} -f "$${tmp_f}";					\
 	done
 
 ######################################################################
